Introduce ed25510 bindings
