+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "sr25519crust"
-version = "0.1.0"
-authors = ["Bogdan Vaneev <warchantua@gmail.com>"]
+version = "0.2.0"
+authors = ["Bogdan Vaneev <warchantua@gmail.com>", "Vladimir Shcherba <abrehchs@gmail.com>"]
 edition = "2018"
 build = "build.rs"
 
@@ -12,10 +12,12 @@ crate-type = ["cdylib", "staticlib"]
 
 [dependencies]
 schnorrkel = { version="0.9.1" }
+ed25519-dalek = { version="1.0.0" }
+rand_chacha = "0.2.2"
 
 [dev-dependencies]
 hex-literal = "0.3.1"
-rand = "0.7.0"
+rand = "0.7"
 
 [build-dependencies]
 cbindgen = "0.14.0"
