+++ README.md
@@ -34,7 +34,7 @@ Header with comments will be generated in `build/include/sr25519/sr25519.h`.
 
 ## Examples
 
-- [Keypair derivation](./test/derive.cpp)
-- [Sign/Verify](./test/ds.cpp)
-- [Keypair from seed](./test/keypair_from_seed.cpp)
-- [Verifiable random function](./test/vrf.cpp)
+- [Keypair derivation](test/sr25519/derive.cpp)
+- [Sign/Verify](test/sr25519/ds.cpp)
+- [Keypair from seed](test/sr25519/keypair_from_seed.cpp)
+- [Verifiable random function](test/sr25519/vrf.cpp)
