+++ build.rs
@@ -1,8 +1,13 @@
-use std::path::Path;
-
-extern crate cbindgen;
-use cbindgen::generate;
+use std::env;
+use cbindgen::Config;
 
 fn main() {
-    generate(Path::new(".")).unwrap().write_to_file(Path::new("./include/sr25519/sr25519.h"));
+    let crate_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
+
+    cbindgen::Builder::new()
+        .with_crate(crate_dir)
+        .with_config(Config::from_file("./cbindgen.toml").expect("Parsing config failed"))
+        .generate()
+        .expect("Unable to generate bindings")
+        .write_to_file("./include/_25519/_25519.h");
 }
