+++ rust.cmake
@@ -11,20 +11,20 @@ endif ()
 
 
 if (BUILD_SHARED_LIBS)
-  set(lib ${path_prefix}/${CMAKE_SHARED_LIBRARY_PREFIX}sr25519crust${CMAKE_SHARED_LIBRARY_SUFFIX})
+  set(lib ${path_prefix}/${CMAKE_SHARED_LIBRARY_PREFIX}_25519crust${CMAKE_SHARED_LIBRARY_SUFFIX})
 else ()
-  set(lib ${path_prefix}/${CMAKE_STATIC_LIBRARY_PREFIX}sr25519crust${CMAKE_STATIC_LIBRARY_SUFFIX})
+  set(lib ${path_prefix}/${CMAKE_STATIC_LIBRARY_PREFIX}_25519crust${CMAKE_STATIC_LIBRARY_SUFFIX})
 endif ()
-message(STATUS "[sr25519] library: ${lib}")
+message(STATUS "[_25519] library: ${lib}")
 
 
 set(include_path ${CMAKE_CURRENT_SOURCE_DIR}/include)
-set(sr25519_h_dir ${include_path}/sr25519)
+set(_25519_h_dir ${include_path}/_25519)
 
 ### setup install task
 include(GNUInstallDirs)
 function(ifd_install what where)
-  message(STATUS "[sr25519] '${what}' will be installed to '${where}'")
+  message(STATUS "[_25519] '${what}' will be installed to '${where}'")
   if (IS_DIRECTORY ${what})
     install(
         DIRECTORY ${what}
@@ -39,10 +39,10 @@ function(ifd_install what where)
 endfunction()
 
 
-ifd_install(${sr25519_h_dir} ${CMAKE_INSTALL_INCLUDEDIR})
+ifd_install(${_25519_h_dir} ${CMAKE_INSTALL_INCLUDEDIR})
 ifd_install(${lib} ${CMAKE_INSTALL_LIBDIR})
-ifd_install(${CMAKE_SOURCE_DIR}/sr25519Config.cmake ${CMAKE_INSTALL_LIBDIR}/cmake/sr25519)
-ifd_install(${CMAKE_SOURCE_DIR}/sr25519Config-noconfig.cmake ${CMAKE_INSTALL_LIBDIR}/cmake/sr25519)
+ifd_install(${CMAKE_SOURCE_DIR}/_25519Config.cmake ${CMAKE_INSTALL_LIBDIR}/cmake/_25519)
+ifd_install(${CMAKE_SOURCE_DIR}/_25519Config-noconfig.cmake ${CMAKE_INSTALL_LIBDIR}/cmake/_25519)
 
 
 ### setup tasks
@@ -53,22 +53,22 @@ add_custom_target(
     WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
 )
 
-add_library(sr25519 STATIC IMPORTED GLOBAL)
+add_library(_25519 STATIC IMPORTED GLOBAL)
 
 # if we build static lib
 if (NOT BUILD_SHARED_LIBS)
   if (APPLE)
     # on apple we need to link Security
     find_library(Security Security)
-    find_package_handle_standard_args(sr25519
+    find_package_handle_standard_args(_25519
         REQUIRED_VARS Security
         )
-    set_target_properties(sr25519 PROPERTIES
+    set_target_properties(_25519 PROPERTIES
         INTERFACE_LINK_LIBRARIES ${Security}
         )
   elseif (UNIX)
     # on Linux we need to link pthread
-    target_link_libraries(sr25519 INTERFACE
+    target_link_libraries(_25519 INTERFACE
         pthread
         -Wl,--no-as-needed
         dl
@@ -78,13 +78,13 @@ if (NOT BUILD_SHARED_LIBS)
   endif ()
 endif ()
 
-set_target_properties(sr25519 PROPERTIES
+set_target_properties(_25519 PROPERTIES
     INTERFACE_INCLUDE_DIRECTORIES ${include_path}
     IMPORTED_LOCATION ${lib}
     )
-add_dependencies(sr25519 cargo_build)
+add_dependencies(_25519 cargo_build)
 
-file(MAKE_DIRECTORY ${sr25519_h_dir})
+file(MAKE_DIRECTORY ${_25519_h_dir})
 
 
 ### add tests
