+++ test/CMakeLists.txt
@@ -4,11 +4,14 @@ find_package(GTest CONFIG REQUIRED)
 find_package(GMock CONFIG REQUIRED)
 
 addtest(sr25519_test
-    ds.cpp
-    derive.cpp
-    keypair_from_seed.cpp
-    vrf.cpp
+        sr25519/ds.cpp
+        sr25519/derive.cpp
+        sr25519/keypair_from_seed.cpp
+        sr25519/vrf.cpp
     )
 target_link_libraries(sr25519_test
     sr25519
     )
+
+addtest(ed25519_test)
+target_link_libraries()
